Refactor pathgrid indices to use size_t instead of int

See merge request OpenMW/openmw!2900

+++ apps/opencs/model/tools/pathgridcheck.hpp
@@ -23,7 +23,7 @@ namespace CSMTools
     struct Point
     {
         unsigned char mConnectionNum;
-        std::vector<int> mOtherIndex;
+        std::vector<size_t> mOtherIndex;
         Point()
             : mConnectionNum(0)
             , mOtherIndex(0)
